# frontend/app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TG Bot Admin UI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/recruiters": {
      "get": {
        "tags": [
          "recruiters"
        ],
        "summary": "Recruiters List",
        "operationId": "recruiters_list_recruiters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/recruiters/new": {
      "get": {
        "tags": [
          "recruiters"
        ],
        "summary": "Recruiters New",
        "operationId": "recruiters_new_recruiters_new_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/recruiters/create": {
      "post": {
        "tags": [
          "recruiters"
        ],
        "summary": "Recruiters Create",
        "operationId": "recruiters_create_recruiters_create_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_recruiters_create_recruiters_create_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/recruiters/{rec_id}/edit": {
      "get": {
        "tags": [
          "recruiters"
        ],
        "summary": "Recruiters Edit",
        "operationId": "recruiters_edit_recruiters__rec_id__edit_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "rec_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rec Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recruiters/{rec_id}/update": {
      "post": {
        "tags": [
          "recruiters"
        ],
        "summary": "Recruiters Update",
        "operationId": "recruiters_update_recruiters__rec_id__update_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "rec_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rec Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_recruiters_update_recruiters__rec_id__update_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recruiters/{rec_id}/delete": {
      "post": {
        "tags": [
          "recruiters"
        ],
        "summary": "Recruiters Delete",
        "operationId": "recruiters_delete_recruiters__rec_id__delete_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "rec_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rec Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recruiters": {
      "get": {
        "tags": [
          "api"
        ],
        "summary": "Api Recruiters",
        "operationId": "api_recruiters_api_recruiters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Api Create Recruiter",
        "operationId": "api_create_recruiter_api_recruiters_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/recruiters/{recruiter_id}": {
      "put": {
        "tags": [
          "api"
        ],
        "summary": "Api Update Recruiter",
        "operationId": "api_update_recruiter_api_recruiters__recruiter_id__put",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "recruiter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recruiter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_recruiters_create_recruiters_create_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tz": {
            "type": "string",
            "title": "Tz",
            "default": "Europe/Moscow"
          },
          "telemost": {
            "type": "string",
            "title": "Telemost",
            "default": ""
          },
          "tg_chat_id": {
            "type": "string",
            "title": "Tg Chat Id",
            "default": ""
          },
          "active": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "cities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cities",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Body_recruiters_create_recruiters_create_post"
      },
      "Body_recruiters_update_recruiters__rec_id__update_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tz": {
            "type": "string",
            "title": "Tz",
            "default": "Europe/Moscow"
          },
          "telemost": {
            "type": "string",
            "title": "Telemost",
            "default": ""
          },
          "tg_chat_id": {
            "type": "string",
            "title": "Tg Chat Id",
            "default": ""
          },
          "active": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "cities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cities",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Body_recruiters_update_recruiters__rec_id__update_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}